Add service_mesh marker (#472) (#521)
Added the `service_mesh` marker to the `pytest.ini` file for tests
requiring the service mesh operator.
Replaced the incorrect `special_infra` marker with
`service_mesh` in the relevant test file. The `special_infra` marker
was replaced because the service mesh operator does not require any
special infrastructure or node-level configurations, making it
inaccurate for tests related to the service mesh.

Signed-off-by: Shahaf Bahar <sbahar@redhat.com>
Co-authored-by: vsibirsk <57763370+vsibirsk@users.noreply.github.com>