[4.18] Mark network tests that don't require node's additional interfaces
* Mark network tests that don't require node's additional interfaces

Acceptance tests for new architectures, platforms etc. shouldn't depend on having extra
NICs on the cluster nodes, except for the primary interface. In order to have satisfactory
coveage in these tests, we mark all the these tests with the a new marker, which can and
should be used for these types of tests.

* Skip multi NIC sanity check if running with single_nic marker