[cnv-4.18] backport localnet test and its dependencies (#944)
* Utilities - Add NodeNetworkConfigurationPolicy resource to libs/net (#218)

- As part of the initiative to create simpler resources and reduce
dependency, this PR introduces a new NNCP resource in the network's team libs
directory.
- The new NNCP is used in an existing test - test_linux_bridge_functionality.

Signed-off-by: Anat Wax <awax@redhat.com>

* tests, net, libs, nncp: On update/patch, assure correct status update (#733)

* tests, network: Cover tests.network.libs with mypy validation

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs, nncp: On update/patch, assure correct status update

When mutating an existing NNCP object, it is common to wait for a
condition status of success (condition type "Available").
However, it has been observed or bare-metal clusters that the observed
condition is the previous one and is not reflecting the actual state
after the spec update/patch.

In order to handle these cases, the condition timestamp is used to
assure it has been updated post the change.

Other alternatives have been found unsafe, like the option of waiting
for an intermediate state (e.g. "progressing"). It is unsafe because the
state is in transition and that can be quicker than the sampling
performed by the test tools.

The current NNCP object updates the spec only when performing the
clean-up step (on teardown). At that step the previous created
interfaces are being removed.
Any future exposure of update/patch requires this technique to assure
successful operation.

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs, nncp: Relocate the resource patching at clean_up

The clean_up method is used to perform teardown operations for any
resource. For the NNCP it is used to revert the interfaces additions
(by deleting them). It makes sense to place the update/patch call at the
cleanup upper level so any other future changes will be able to re-use
the same single patch operation.

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs, nncp: Drop the condition class and fix func signature

Signed-off-by: Edward Haas <edwardh@redhat.com>

---------

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs: Introduce network cloud-init data structure (#754)

* tests, net, libs: Introduce network cloud-init data structure

Introduce a new data structure for network cloud-init.
It is aimed to replace the existing dict-based configuration.

Usage example:
```
from tests.network.libs import cloudinit

nd = cloudinit.NetworkData(ethernets={
    "eth0": EthernetDevice(
        dhcp4=False,
        addresses=["1.1.1.1/24", "d10:0:2::2"],
        gateway6="d10:0:2::1",
    )
})

nd_yaml = yaml.safe_dump(cloudinit.todict(netdata=nd), sort_keys=False)
cloudinit_dict = cloudinit.cloudinit(netdata=nd)
```

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, bond: First use of the new structured cloud-init network

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs: Introduce 'apimachinery' module

The new module is aimed to contain API related helpers.
First candidate: the dict normalization which is used to translate a
dataclass to a dict.

Signed-off-by: Edward Haas <edwardh@redhat.com>

---------

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs, nncp:  Assure NNCP resource exists, status & conditions absense is treated correctly. (#876)

* tests, net, libs, nncp: Assure resource is waited for on deployment

Currently, the default resource deployment is not waiting for the
resource to appear. We are interested for the NNCP resources to assure
that the resource exists before we continue.

This change assures that the `self.instance` exists and with it the
`status`.

Signed-off-by: Edward Haas <edwardh@redhat.com>

* tests, net, libs, nncp: Do not assume NNCP 'status' & 'conditions' exist

On NNCP resource creation, the status and/or conditions may not yet be
populated. Assure such scenarios are treated.

Signed-off-by: Edward Haas <edwardh@redhat.com>

---------

Signed-off-by: Edward Haas <edwardh@redhat.com>

* Add test for connectivity between localnet VMs (#464)

* Ensure boot disk is first in list of VM disks and volumes

The default boot image must be the first disk for correct VM behavior.
Changed disk and volume insertion to use insert instead of append
so the boot image appears first in the disks and volumes lists.

Signed-off-by: Shahaf Bahar <sbahar@redhat.com>

* Add test for connectivity between localnet VMs

Introduced a new test to verify connectivity between VMs on the
localnet network. This test ensures that VMs within the localnet
network have connectivity.

Signed-off-by: Shahaf Bahar <sbahar@redhat.com>

---------

Signed-off-by: Shahaf Bahar <sbahar@redhat.com>

---------

Signed-off-by: Anat Wax <awax@redhat.com>
Signed-off-by: Edward Haas <edwardh@redhat.com>
Signed-off-by: Shahaf Bahar <sbahar@redhat.com>
Co-authored-by: ‪Anat Wax <46938838+Anatw@users.noreply.github.com>
Co-authored-by: Edward Haas <edwardh@redhat.com>