github-action, component-builder: Change current workflow to check-only (#1031)
* github-action, component-builder: Change current workflow to check-only

The existing workflow to build the Fedora image has a flow: The event
is 'pull_request_target' which runs in the context of the target branch
and therefore does not include the changes a PR introduces.

In order to run the changes in the PR, the event must be changed to
'pull_request'. However, this event is restricting access to the secrets
which are needed to 1) generate the password and 2) push the image to
the registry (quay).

The solution is to convert the current workflow to a build-only check,
i.e. to check that the image can be built, but without publishing the
image to the registry.
The publish itself to the registry is kept by introducing another
workflow with a 'push' event. It is triggered post merge to persist the
image.
At this stage, it is still pushing to the staging repo but in the future
it should be change to the production one.

Follow up changes should include making the image available for
consumption by users (e.g. as an artifact) so it can be validated.

Signed-off-by: Edward Haas <edwardh@redhat.com>

* gh action, component-builder: Upload the generated image as an artifact

PR authors can now access the created container image and use to for its
verification.

After downloading the tarball, load it locally:
`podman load -i download/fedora-image.tar`

Signed-off-by: Edward Haas <edwardh@redhat.com>

---------

Signed-off-by: Edward Haas <edwardh@redhat.com>