tests, net, localnet: Refactor the fixtures to express better the action (#1052)
The fixture that is running the VMs is renamed to express that it does
exactly that.
The fixture that is migrating a VM is changed to depend on an additional
fixture that assures connectivity.

The change expresses better that there is no meaning to "client" and
"server" VMs.

In addition, constants have been moved in order to use them outside the
fixtures files.

Signed-off-by: Edward Haas <edwardh@redhat.com>