net tests, hotplug: use unprivileged client for migration (#1247)
jira-ticket: https://issues.redhat.com/browse/CNV-63868

Starting with release 4.19, VM migration can be performed by either
a cluster admin or a namespace admin with an assigned role -
kubevirt.io:migrate. In hot-plug tests, VM migration is performed
by the admin cluster, which is not entirely correct from the point of
view of the user scenario. It is more correct to grant the necessary
rights to the namespace admin and migrate from the namespace admin.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>