[CNV-64119]Fix test_subscription_channels.test_cnv_subscription_channel (#1464)
While testing stable releases, we might have both stable & candidate
channels on the same packagemanifest version:

$ oc get packagemanifest kubevirt-hyperconverged \
  -n openshift-marketplace -o json | jq -r ".status.channels[] |\
  select(.currentCSVDesc.version==\"4.19.1\") | .name"
candidate
stable

The behaviour before this change was to take the first one, which was
mismatching with the channel set on the subscription ("stable"). The
test was failing in this case.

This change is checking that the subscription channel is present in the
list of available channels in the packagemanifest.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>