Exceptional skip of IPv6 scenario on non-IPv4-single-stack cluster
Although we avoid programmatic skips, this one is an exception where we must skip.
The reason is that this test, which is a regression test of an old bug, is
required to run only on a single-stack IPv4 cluster. We currenly don't have a way
of marking a cluster as single-stack, and in addition - it wouldn't make sense to
maintain a dedicated test lane just for this one test.