Attempt to fix IP assignment flakiness
We occasionally see network tests failing on setup due to IP not assigned to an interface
during the grace timeout we grant to it.
I tried debugging the logs in such failure, but didn't find any error or info regarding
that interface - neither in the guest, the VMI or the pod.
Therefore it seems that it might be worth waiting a bit more for the IP to get updated.
Another reason is that if the problem is not with a insufficient timeout, then we will
keep hitting this failure, which will bring to the conclusion that the problem is somewhere
else, thus requires further debugging.