tests, net, sanity: Add IPv6 family stack validation (#2177)
Tests which are marked with the `ipv6` marker are expecting the cluster
to support IPv6. If it does not support IPv6, the infra is considered
incorrect and therefore no tests should execute.

This change refactors the IPv4 sanity check and extends it to cover IPv6
as well.

Follow up changes should convert usages of
`fail_if_not_ipv6_supported_cluster` to a test marker.

Signed-off-by: Edward Haas <edwardh@redhat.com>