net, traffic_generator: improve naming + retry decorator using (#2389)
Client\Server looks quite common and such naming does not reflect
the real intention. The usage traffic_generator.Client does not
solve the problem as well. It's crucial to explicitly point out that
we work with Tcp traffic and the classes implement TCP client and
TCP server in reality. And this clarity is significant for tests also.

The commit also includes replacing TimeoutSampler with retry decorator,
that leads a little refactoring for classes to align the logic.
Retry-enabled `_ensure_is_running()` method is added to `__enter__()`.
PodTcpClient already uses a similar implementation. We need to make
sure once when entering into the context manager that the client\server
process is running, then it is enough to use the atomic method
`is_running()` without retry to determine the status of the process.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>