[4.18] Storage: remove unneeded tests and checks in import and hpp modules (#2667) (#2683)
Manual backports:
https://github.com/RedHatQE/openshift-virtualization-tests/pull/2574 
https://github.com/RedHatQE/openshift-virtualization-tests/pull/1845 
https://github.com/RedHatQE/openshift-virtualization-tests/pull/1856

<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged If
the task is not tracked by a Jira ticket, just write "NONE". -->

---------

##### Short description:

##### More details:

##### What this PR does / why we need it:

##### Which issue(s) this PR fixes:

##### Special notes for reviewer:

##### jira-ticket:
<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged
If the task is not tracked by a Jira ticket, just write "NONE".
-->

Co-authored-by: dalia-frank <65129017+dalia-frank@users.noreply.github.com>