[Storage] Add DV list and get permissions in restricted_namespace_cloning tests (#2693)
##### Short description:
The client was added to the `running_vm()` when it's waiting for DV
succees:
`DataVolume(name=dv_name, namespace=_vm.namespace,
**client=_vm.client**).wait_for_dv_success()`

So if the user doesn't have permissions to access the DV status, the
tests are failing.

##### More details:
Jira: https://issues.redhat.com/browse/CNV-72460

##### What this PR does / why we need it:

##### Which issue(s) this PR fixes:

##### Special notes for reviewer:

##### jira-ticket:
<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged
If the task is not tracked by a Jira ticket, just write "NONE".
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Tests**
* Extended test coverage for VM cloning scenarios with data volumes in
restricted namespaces
* Enhanced permission validation testing for data volume destination
access
* Improved test parameterization to cover additional namespace
restriction scenarios

<!-- end of auto-generated comment: release notes by coderabbit.ai -->