net, sriov: Remove redundant sriov policy and fixtures (#2809)
In current setup, SR-IOV tests are passing successfully only on a BM,
which has a special setup and multiple physical SR-IOV interfaces. When
running sriov tests on other BMs, the tests fail due to:
- Vlan interface and SR-IOV unused interface don't match causing the
setup of tests to fail. VLAN for workloads is configured in the NAD (via
SriovNetwork), not on the node networking.
- Adding an extra policy `test-sriov-on-vlan-policy` on bm with only one
physical sriov intefrace - leads to creating 2 policies on one physical
sriov interface, vm3 and vm4 are stuck on Scheduling and tests fail.

Fixes:
These fixes will enable to run sriov tests on every BM with sriov
support.
- Remove redundant fixtures: There's no need in a dedicated policy which
`sriov_with_vlan_node_policy` produced, and no need to select a
dedicated interface for VLAN. All tests can and should use the same
configuration policy, which is implemented in `sriov_node_policy`.
- Use 1 policy for all vms: `sriov_network_vlan` is now using
`sriov_node_policy`.

This PR was split from this:
https://github.com/RedHatQE/openshift-virtualization-tests/pull/2760

##### jira-ticket: https://issues.redhat.com/browse/CNV-72408

<!-- full-ticket-url needs to be provided. This would add a link to the
pull request to the jira and close it when the pull request is merged
If the task is not tracked by a Jira ticket, just write "NONE".
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Tests**
* Updated network testing fixtures and dependencies to align with
current test structure.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>

<!-- end of auto-generated comment: release notes by coderabbit.ai -->